requirements.txt
pandas>=1.3.0
numpy>=1.19.0
seaborn>=0.11.0
matplotlib>=3.4.0
python-pptx>=0.6.0
jupyter>=1.0.0
